requirements.txt
absl-py==1.0.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.12
cloudpickle==2.0.0
cycler==0.11.0
cytoolz==0.11.0
dask==2021.3.0
dataclasses==0.8
decorator==5.1.1
filelock==3.4.1
future==0.18.2
gast==0.3.3
gdown==3.11.1
gluoncv==0.7.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.44.0
h5py==2.10.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.3
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mxnet==1.6.0
networkx==2.5
numpy==1.18.2
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.2.0.34
opencv-python==4.5.5.64
opt-einsum==3.3.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip==21.2.2
portalocker==2.4.0
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.1.1
pywebview==3.2
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.15.0
scipy==1.4.1
setuptools==58.0.4
six==1.16.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.1
tensorflow-estimator==2.2.0
termcolor==1.1.0
toolz==0.11.2
torch==1.5.0
torchvision==0.2.2.post3
tornado==6.1
tqdm==4.43.0
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.14.0
zipp==3.6.0
